Moved check attribute type method to Util.
